Add dynamic OHLC chart
