+++ .storybook/TypeSwitchableChart.js
@@ -1,0 +1,38 @@
+import React from 'react';
+import BinaryChart from '../src/BinaryChart';
+import api from './ApiSingleton';
+
+export default class TypeSwitchChart extends React.Component {
+    constructor(props) {
+        super(props);
+        this.state = {
+            ticks: [],
+            type: 'area',
+        };
+    }
+
+    changeType(t) {
+        switch (t) {
+            case 'tick': api.getTickHistory('R_100').then(r => {
+                const ticks = r.history.times.map((t, idx) => {
+                    const quote = r.history.prices[idx];
+                    return { epoch: +t, quote: +quote };
+                });
+                this.setState({ ticks, type: 'area' });
+            });
+                break;
+            case 'candlestick': api.getCandles('R_100').then(r => {
+                this.setState({ type: 'candlestick', ticks: convertEpochToMS(rawData)});
+            });
+                break;
+            default: return;
+        }
+    }
+
+    render() {
+        const { ticks, type } = this.state;
+        return (
+            <BinaryChart type={type} ticks={ticks} typeChange={t => this.changeType(t)} />
+        )
+    }
+}
