+++ .storybook/ohlc.js
@@ -1,7 +1,8 @@
 import React from 'react';
 import { storiesOf } from '@kadira/storybook';
 import BinaryChart from '../src/BinaryChart';
-import api from './ApiSingleton';
+import TypeSwitchChart from './TypeSwitchableChart';
+import DynamicOHLCChart from './DynamicOHLCChart';
 
 const rawData =  [
     {
@@ -75,42 +76,7 @@ const rawData =  [
         low: "678.7806"
     }
 ];
-const convertEpochToMS = dataArr => dataArr.map(d => Object.assign(d, { epoch: d.epoch * 1000 }));
-
-class TypeSwitchChart extends React.Component {
-    constructor(props) {
-        super(props);
-        this.state = {
-            ticks: [],
-            type: 'area',
-        };
-    }
-
-    changeType(t) {
-        switch (t) {
-            case 'tick': api.getTickHistory('R_100').then(r => {
-                const ticks = r.history.times.map((t, idx) => {
-                    const quote = r.history.prices[idx];
-                    return { epoch: +t, quote: +quote };
-                });
-                this.setState({ ticks, type: 'area' });
-            });
-                break;
-            case 'candlestick': api.getCandles('R_100').then(r => {
-                this.setState({ type: 'candlestick', ticks: convertEpochToMS(rawData)});
-            });
-                break;
-            default: return;
-        }
-    }
-
-    render() {
-        const { ticks, type } = this.state;
-        return (
-            <BinaryChart type={type} ticks={ticks} typeChange={t => this.changeType(t)} />
-        )
-    }
-}
+export const convertEpochToMS = dataArr => dataArr.map(d => Object.assign(d, { epoch: d.epoch * 1000 }));
 
 storiesOf('OHLC', module)
     .add('Simple OHLC chart', () =>
@@ -122,4 +88,7 @@ storiesOf('OHLC', module)
             <TypeSwitchChart />
         );
     })
+    .add('OHLC streaming', () =>
+        <DynamicOHLCChart />
+    )
 ;
